[EDDI] Fixes how addConsistencyChecks handles the new verification basic block
Fixes HEAPLab/ASPIS#2.
When creating a VerificationBB between the exception handling and the call of __clang_call_terminate, the invoke landing pad address was substituted with the verificationBB address instead of the exception handler address. This lead to the invoke unwind label jumping after the exception handling part, leading to the "The unwind destination does not have an exception handling instruction!" error.