feat(RACFED): add pass for RACFED implementation
LLVM pass implementing Random Additive Control Flow Error Detection (RACFED).
Original algorithm by Vankeirsbilck et Al. (DOI: 10.1007/978-3-319-99130-6_15)

Random Additive Control Flow Error Detection is a method that is
able to catch subsequent instruction skips (>= 1).
This method, similarly to RASM, assignes statically to each Basic Block (BB) a
signature.
It then consists in updating a variable (in our case global), named
runtime-signature (in our case signature), that keeps track of the instructions
run and when met with jump instructions (branch/return) it checks whether the
runtime-signature is compatible with the signature statically assigned
to the block.
It is stronger than RASM because every after every instruction the
runtime-signature is updated.

This implementation of RACFED tries to map as close as possible the
RACFED algorithm.
Shortcomings of our implementation are additional overhead due to how
functions are handled (they are not considered endpoints for BBs).

Co-authored-by: Gab-San (Gabriele Santandrea) <gabriele.santandrea@tutamail.com>